+++ CHANGELOG.md
@@ -17,6 +17,7 @@ before)
 - Download of the code now is possible
 - When running all above the screen jumps to the receipt of the running chunk
 - Run all above now mark all the chunks above as running
+- Login using SSH Identity file
 
 ### Internal
 - Protects the user password in the session store